refactor llm config stuff
